fix xseg-tool rnd functions to check for error on submit
also changeg type of xqindex to uint64_t

+++ xseg/xtypes/xq.h
@@ -3,7 +3,7 @@
 
 #include <sys/util.h>
 
-typedef uint32_t xqindex;
+typedef uint64_t xqindex;
 
 #include "xlock.h"
